requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
mandrill==1.0.59
email-validator==2.1.0.post1
cryptography==41.0.5
gunicorn==21.2.0
